@@ -2,10 +2,13 @@
     "name": "qs",
     "description": "A querystring parser that supports nesting and arrays, with a depth limit",
     "homepage": "https://github.com/ljharb/qs",
-    "version": "6.5.2",
+    "version": "6.11.0",
     "repository": {
         "type": "git",
         "url": "https://github.com/ljharb/qs.git"
+    },
+    "funding": {
+        "url": "https://github.com/sponsors/ljharb"
     },
     "main": "lib/index.js",
     "contributors": [
@@ -17,36 +20,58 @@
     ],
     "keywords": [
         "querystring",
-        "qs"
+        "qs",
+        "query",
+        "url",
+        "parse",
+        "stringify"
     ],
     "engines": {
         "node": ">=0.6"
     },
-    "dependencies": {},
+    "dependencies": {
+        "side-channel": "^1.0.4"
+    },
     "devDependencies": {
-        "@ljharb/eslint-config": "^12.2.1",
-        "browserify": "^16.2.0",
-        "covert": "^1.1.0",
-        "editorconfig-tools": "^0.1.1",
-        "eslint": "^4.19.1",
-        "evalmd": "^0.0.17",
-        "iconv-lite": "^0.4.21",
-        "mkdirp": "^0.5.1",
+        "@ljharb/eslint-config": "^21.0.0",
+        "aud": "^2.0.0",
+        "browserify": "^16.5.2",
+        "eclint": "^2.8.1",
+        "eslint": "=8.8.0",
+        "evalmd": "^0.0.19",
+        "for-each": "^0.3.3",
+        "has-symbols": "^1.0.3",
+        "iconv-lite": "^0.5.1",
+        "in-publish": "^2.0.1",
+        "mkdirp": "^0.5.5",
+        "npmignore": "^0.3.0",
+        "nyc": "^10.3.2",
+        "object-inspect": "^1.12.2",
         "qs-iconv": "^1.0.4",
-        "safe-publish-latest": "^1.1.1",
+        "safe-publish-latest": "^2.0.0",
         "safer-buffer": "^2.1.2",
-        "tape": "^4.9.0"
+        "tape": "^5.5.3"
     },
     "scripts": {
-        "prepublish": "safe-publish-latest && npm run dist",
+        "prepack": "npmignore --auto --commentLines=autogenerated",
+        "prepublishOnly": "safe-publish-latest && npm run dist",
+        "prepublish": "not-in-publish || npm run prepublishOnly",
         "pretest": "npm run --silent readme && npm run --silent lint",
-        "test": "npm run --silent coverage",
-        "tests-only": "node test",
+        "test": "npm run tests-only",
+        "tests-only": "nyc tape 'test/**/*.js'",
+        "posttest": "aud --production",
         "readme": "evalmd README.md",
-        "prelint": "editorconfig-tools check * lib/* test/*",
-        "lint": "eslint lib/*.js test/*.js",
-        "coverage": "covert test",
+        "postlint": "eclint check $(git ls-files | xargs find 2> /dev/null | grep -vE 'node_modules|\\.git' | grep -v dist/)",
+        "lint": "eslint --ext=js,mjs .",
         "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js"
     },
-    "license": "BSD-3-Clause"
+    "license": "BSD-3-Clause",
+    "publishConfig": {
+        "ignore": [
+            "!dist/*",
+            "bower.json",
+            "component.json",
+            ".github/workflows"
+        ]
+    }
 }
