@@ -1,40 +1,38 @@
-const token = require("../app/token.js");
-const Token = new token();
-const User = require("../models").User;
-var moment = require("moment");
+import User from "../models/User.js";
 
-module.exports = {
-  index(req, res) {
-    const decoded = Token.decode(req.get("Authorization"));
-    User.findOne({ where: { id: req.params.userId || decoded.user_id } })
-      .then((user) => {
-        res.status(200).send({
-          bio: user.bio,
-          nick: user.nick,
-          dateOfBirth: moment(user.dateOfBirth).format("DD-MM-YYYY"),
-          name: `${user.firstName} ${user.lastName}`,
-          avatar: user.avatar,
-          username: user.username,
-        });
-      })
-      .catch((error) => {
-        res.status(400).send({ message: req.t("error") });
-      });
-  },
-  update(req, res) {
-    const decoded = Token.decode(req.get("Authorization"));
-    User.update(
-      {
-        bio: req.body.bio,
-        username: req.body.username,
-      },
-      { where: { id: decoded.user_id } }
-    )
-      .then(() => {
-        res.status(200).send({ message: req.t("user_updated") });
-      })
-      .catch((error) => {
-        res.status(400).send({ message: req.t("error") });
-      });
-  },
-};
+export const updateUser = async (req,res,next)=>{
+  try {
+    const updatedUser = await User.findByIdAndUpdate(
+      req.params.id,
+      { $set: req.body },
+      { new: true }
+    );
+    res.status(200).json(updatedUser);
+  } catch (err) {
+    next(err);
+  }
+}
+export const deleteUser = async (req,res,next)=>{
+  try {
+    await User.findByIdAndDelete(req.params.id);
+    res.status(200).json("User has been deleted.");
+  } catch (err) {
+    next(err);
+  }
+}
+export const getUser = async (req,res,next)=>{
+  try {
+    const user = await User.findById(req.params.id);
+    res.status(200).json(user);
+  } catch (err) {
+    next(err);
+  }
+}
+export const getUsers = async (req,res,next)=>{
+  try {
+    const users = await User.find();
+    res.status(200).json(users);
+  } catch (err) {
+    next(err);
+  }
+}
